finishing devtalles section 6

+++ devtalles/nextjs14/03-admin-todos/src/app/api/todos/route.ts
@@ -1,10 +1,16 @@
+import { auth } from '@/auth'
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup'
 
 
 // funcion para listar todos
-export async function GET(request: Request) { 
+export async function GET(request: Request) {
+  const session = await auth()
+  if(!session?.user?.id)
+    return NextResponse.json({message: 'User not found'}, {status: 400})
+  
+  
   const { searchParams } = new URL(request.url)
 
   // Check if the take parameter is a number
@@ -20,6 +26,7 @@ export async function GET(request: Request) {
   const todos = await prisma.todo.findMany({
     take: parseInt(take),
     skip: parseInt(skip),
+    where: { userId: session.user.id },
   })
   return NextResponse.json({
     todos
@@ -34,11 +41,19 @@ const postSchema = yup.object({
 // funcion para crear un solo todo
 export async function POST(request: Request) {
   try {
+    const session = await auth()
+    if(!session?.user?.id)
+      return NextResponse.json({message: 'User not found'}, {status: 400})
+    
+    
     const body = await postSchema.validate(await request.json())
+    
+    
     const todo = await prisma.todo.create({
       data: {
         description: body.description,
-        completed: body.completed
+        completed: body.completed,
+        userId: session.user.id
       }
     })
     return NextResponse.json(todo)
@@ -53,13 +68,17 @@ export async function POST(request: Request) {
 
 // funcion para eliminar todos los todos completados
 export async function DELETE(request: Request) {
+  const session = await auth()
+  if(!session?.user?.id)
+    return NextResponse.json({message: 'User not found'}, {status: 400})
+  
   const completedTodos = await prisma.todo.findMany({
-    where: { completed: true }
+    where: { completed: true, userId: session.user.id }
   })
   if(completedTodos.length === 0) return NextResponse.json({message: 'No completed todos'})
   
   await prisma.todo.deleteMany({
-    where: { completed: true }
+    where: { completed: true, userId: session.user.id }
   })
   return NextResponse.json({completedTodos})
 }
